requirements.txt
Flask==2.2.2
line-bot-sdk==2.3.0
python-dotenv
gunicorn